refactor(llm-agent): move CircuitBreaker, FallbackRag from llm-agent-server
Move resilience module (CircuitBreaker, CircuitBreakerLlm, CircuitBreakerEmbedder,
FallbackRag) from llm-agent-server into llm-agent to enable building the package
independently. Tests move with the implementation; server re-exports from llm-agent.
Added makeRag test helper to llm-agent testing module to support moved resilience tests.
