+++ packages/llm-agent/src/resilience/index.ts
@@ -1,0 +1,8 @@
+export {
+  CircuitBreaker,
+  type CircuitBreakerConfig,
+  type CircuitState,
+} from './circuit-breaker.js';
+export { CircuitBreakerEmbedder } from './circuit-breaker-embedder.js';
+export { CircuitBreakerLlm } from './circuit-breaker-llm.js';
+export { FallbackRag } from './fallback-rag.js';
